removed feed from users service, users from feed service, and removed original combined backend
