+++ udacity-c2-restapi-feed/src/controllers/v0/feed/models/FeedItem.ts
@@ -1,5 +1,4 @@
 import {Table, Column, Model, HasMany, PrimaryKey, CreatedAt, UpdatedAt, ForeignKey} from 'sequelize-typescript';
-import { User } from '../../users/models/User';
 
 @Table
 export class FeedItem extends Model<FeedItem> {
