+++ udacity-c2-restapi-feed/src/controllers/v0/model.index.ts
@@ -1,4 +1,3 @@
 import { FeedItem } from './feed/models/FeedItem';
-import { User } from './users/models/User';
 
-export const V0MODELS = [ FeedItem, User ];
+export const V0MODELS = [ FeedItem ];
