--- udacity-c2-restapi/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "udacity-c2-restapi",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/server.js",
-  "scripts": {
-    "start": "node .",
-    "tsc": "tsc",
-    "dev": "ts-node-dev --respawn --transpileOnly ./src/server.ts",
-    "prod": "tsc && node ./www/server.js",
-    "clean": "rm -rf www/ || true",
-    "build": "npm run clean && tsc && cp -rf src/config www/config && cp .npmrc www/.npmrc && cp package.json www/package.json && cd www && zip -r Archive.zip . && cd ..",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "Gabriel Ruttner",
-  "license": "ISC",
-  "dependencies": {
-    "@types/jsonwebtoken": "^8.3.2",
-    "aws-sdk": "^2.492.0",
-    "bcrypt": "^3.0.6",
-    "body-parser": "^1.19.0",
-    "email-validator": "^2.0.4",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "pg": "^7.11.0",
-    "reflect-metadata": "^0.1.13",
-    "sequelize": "^5.10.0",
-    "sequelize-typescript": "^0.6.11"
-  },
-  "devDependencies": {
-    "@types/bcrypt": "^3.0.0",
-    "@types/bluebird": "^3.5.27",
-    "@types/express": "^4.17.0",
-    "@types/node": "^11.13.17",
-    "@types/sequelize": "^4.28.3",
-    "@types/validator": "^10.11.1",
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "mocha": "^6.1.4",
-    "ts-node-dev": "^1.0.0-pre.40",
-    "tslint": "^5.18.0",
-    "typescript": "^3.5.3"
-  }
-}
