--- udacity-c2-restapi/src/controllers/v0/index.router.ts
@@ -1,14 +1,0 @@
-import { Router, Request, Response } from 'express';
-import { FeedRouter } from './feed/routes/feed.router';
-import { UserRouter } from './users/routes/user.router';
-
-const router: Router = Router();
-
-router.use('/feed', FeedRouter);
-router.use('/users', UserRouter);
-
-router.get('/', async (req: Request, res: Response) => {    
-    res.send(`V0`);
-});
-
-export const IndexRouter: Router = router;
